requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langchain-qdrant
langchain-text-splitters
qdrant-client
sentence-transformers
transformers
fastapi
uvicorn
python-dotenv
pypdf
pydantic
groq
torch
accelerate